SCard Improvements + S2 Issues
TeamView now has the functionality to assign users to projects and remove users from projects
-Some buggy behavior can be observed with this but I think it may have been resolved in other commits made this morning/last night
->Bugs include duplicate projects in listOfAssignedProjects for user

SCardDescription properly shows the priority now. Some of the priorities are still zero, so it is configured to be 0 in that case.

Main View shows assigned items when clicked. (Weird behavior was notices in listbox responsiveness when only 1 listbox was bound to event)

Removed labels for main view and developer home view. Also removed selectedUserName line that wasn't used for anything.

Issues: #122, #116, #111