Developer Home View Start for UX Goals
A developer home view change to bring it inline with #193, only thing missing is the progress bar.

Changelog:
commentsField has been changed to require that masterView be provided to it as the second parameter.

commentFeed has been given a function updateComments() that should be called in your views callback function to allow for comments to update with the db (takes a little while to show up)

Comments now have an entry field that you can put stuff in, click the submit button and it will be uploaded to the DB and changed on the application.

updateFromListOfCommentsObject() has been changed so it takes in a source and something to seach the source for (usually activeProject or currentUser and event.widget.get(tk.ANCHOR)). This is done for the realtime update functionality.

A minor packing change has been made to backlog lists so that if the external list is expanded, the contents will as well.

Fixed comments being sorted from least recent to most: now its from most to least

@TODO Make Comments Scrollable (Will allow for more comments to be displayed, and should allow for all elements in our text field entry to be displayed: right now if there are a lot of comments then the submit button will be cut-off).