created methods to handle loss of internet connection. Tested by toggling airplane mode on laptop.  Function is reasonable.
MasterView passes a loss of connection handler function into a datablock method onConnectionLoss as a separate thread that loops over a connection true flag.  when the flag is false, the loop exits and the handler is executed.

Passes in a MasterView.connectionLossHandler which calls logOut.
If no internet connection on attempted login, application shut's down

Signed-off-by: fallinbryan <fallinbryan@ufl.edu>