Add Button
Put this in this branch because why not.

Altered backlog view to have a + button to add new items, can be disabled with the canAdd [default is True] params (disable on any view where adding would be confusing: tied to sprint, epic or user).

Only on backlog views because they are item only (would need additional checks for normal SLists that are unecessary for a little utility like this)

#199